option to disable file deletion after process

+++ app/Jobs/processXML.php
@@ -58,7 +58,7 @@ class processXML implements ShouldQueue
             $epg->save();
             $z->next('programme');
         }
-        Storage::delete($this->file_name);
+        if(config('app.delete_file')) Storage::delete($this->file_name);
         if(config('app.redis_enabled')) Redis::flushDB();
     }
 }
